- Changed size_t to uint32_t for function and calls to  GetBufferPointerUnsafe as this inconsistency caused  some compilers to fail.

+++ trunk/src/common/net/messages.cpp
@@ -7133,7 +7133,7 @@ psCachedFileMessage::psCachedFileMessage( uint32_t client, const char *pathname,
 psCachedFileMessage::psCachedFileMessage( MsgEntry* me )
 {
 	hash = me->GetStr();
-	size_t size=0;
+	uint32_t size=0;
 	char *ptr = (char *)me->GetBufferPointerUnsafe(size);
 	if (ptr)
 	{
